DEV: Isolate multisite specs
Mixing multisite and standard specs can lead to issues (e.g. when using `fab!`)

Disabled the (upcoming) rubocop rule for two files that have thoroughly tangled both types of specs.
