+++ spec/components/site_setting_extension_spec.rb
@@ -162,17 +162,6 @@ describe SiteSettingExtension do
     end
   end
 
-  describe "multisite" do
-    it "has no db cross talk", type: :multisite do
-      settings.setting(:hello, 1)
-      settings.hello = 100
-
-      test_multisite_connection("second") do
-        expect(settings.hello).to eq(1)
-      end
-    end
-  end
-
   describe "DiscourseEvent" do
     before do
       settings.setting(:test_setting, 1)
