+++ spec/integration/multisite_spec.rb
@@ -49,7 +49,6 @@ describe 'multisite', type: :multisite do
   end
 
   it "should hit correct site elsewise" do
-
     headers = {
       "HTTP_HOST" => "test2.localhost",
       "REQUEST_METHOD" => "GET",
@@ -61,5 +60,4 @@ describe 'multisite', type: :multisite do
     expect(code).to eq(200)
     expect(body.join).to eq("test2.localhost")
   end
-
 end
