+++ spec/lib/backup_restore/shared_examples_for_backup_store.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+# rubocop:disable Discourse/OnlyTopLevelMultisiteSpecs
 
 shared_context "backups" do
   before { create_backups }
