+++ spec/lib/backup_restore/uploads_restorer_spec.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+# rubocop:disable Discourse/OnlyTopLevelMultisiteSpecs
 
 require 'rails_helper'
 require_relative 'shared_context_for_backup_restore'
